vagrant: update net-next to 141
This VM contains the containerd container rumtime that is going to be
used with Kubernetes 1.24 as Kubernetes dropped support for docker.

Due to some changes done in net-next we will need to allow all
ports in the network 192.168.61.0.

Also, switched the container runtime to containerd since that is already
available in K8s 1.23 and we can start using it this version. If we keep
docker, the container images will have to be pulled by the docker container
runtime making the tests slower and eventually timeout.

Signed-off-by: André Martins <andre@cilium.io>